move psi utils to util package
