+++ src/main/java/com/github/stefvanschie/quickskript/psi/literal/PsiNumberLiteral.java
@@ -2,6 +2,7 @@ package com.github.stefvanschie.quickskript.psi.literal;
 
 import com.github.stefvanschie.quickskript.psi.PsiConverter;
 import com.github.stefvanschie.quickskript.psi.PsiElementFactory;
+import com.github.stefvanschie.quickskript.psi.util.PsiPrecomputedHolder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
