+++ src/main/java/com/github/stefvanschie/quickskript/psi/literal/PsiStringLiteral.java
@@ -2,6 +2,7 @@ package com.github.stefvanschie.quickskript.psi.literal;
 
 import com.github.stefvanschie.quickskript.psi.PsiConverter;
 import com.github.stefvanschie.quickskript.psi.PsiElementFactory;
+import com.github.stefvanschie.quickskript.psi.util.PsiPrecomputedHolder;
 import com.github.stefvanschie.quickskript.util.TextMessage;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
